pyproject.toml
[project]
name = "bank"
version = "0.2.0"
description = "Multi-agent Japanese equity intelligence and reporting system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "edinet-xbrl>=0.2.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
